requirements.txt
argparse==1.4.0
grad-cam==1.5.0
imbalanced-learn==0.10.1
imblearn==0.0
imgaug==0.4.0
lifelines==0.27.8
mamba-ssm==1.2.0.post1
markdown==3.4.3
matplotlib==3.7.1
monai==1.2.0
nibabel==5.1.0
numpy==1.24.3
opencv-python==4.8.1.78
openpyxl==3.1.2
pandas==2.0.2
plotly==5.18.0
ptitprince==0.2.7
ptyprocess==0.7.0
pydicom==2.4.0
pynrrd==1.0.0
pyopenssl==23.0.0
pyparsing==3.0.9
pyradiomics==3.1.0
pytorch==2.0.1
pywavelets==1.4.1
pyyaml==6.0.1
sci-palettes==1.0.1
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.11.0
simpleitk==2.3.0
simplenomo==1.0.0
six==1.16.0
stack_data==0.6.2
statannot==0.2.3
statsmodels==0.14.0
sympy==1.11.1
tableone==0.7.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
timm==0.9.12
torchaudio==2.0.2
torchio==0.18.92
torchsampler==0.1.2
torchvision==0.15.2
tqdm==4.65.0

